trends_app/requirements.txt
# Trends App - Core Dependencies
# Python 3.8+ required

# ============================================================================
# Core Flask Dependencies
# ============================================================================
Flask>=2.3.0,<3.0
flask-login>=0.6.0
Werkzeug>=2.3.0

# ============================================================================
# Database
# ============================================================================
psycopg2-binary>=2.9.0

# ============================================================================
# Configuration & Utilities
# ============================================================================
PyYAML>=6.0
Jinja2>=3.1.0

# ============================================================================
# HTTP Requests & API
# ============================================================================
requests>=2.31.0

# ============================================================================
# Data Processing
# ============================================================================
python-dateutil>=2.8.0
deepdiff>=6.3.0

# ============================================================================
# Encryption (AWS KMS support - optional)
# ============================================================================
# Required if using encryption.mode: kms in config
boto3>=1.28.0
cryptography>=41.0.0

# ============================================================================
# Production WSGI Server (Recommended)
# ============================================================================
gunicorn>=21.0.0

# ============================================================================
# Optional: Async Queue Backend (for high-scale deployments)
# ============================================================================
# Uncomment if using submission_mode: async_queue
# celery>=5.3.0
# redis>=5.0.0

# ============================================================================
# Optional: Development/Testing
# ============================================================================
# Uncomment for development
# pytest>=7.4.0
# pytest-flask>=1.2.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0

# ============================================================================
# Version Notes
# ============================================================================
# - Flask 2.3+ required for modern features
# - psycopg2-binary is used instead of psycopg2 for easier installation
#   (includes pre-compiled C libraries)
# - boto3 and cryptography only needed if using AWS KMS encryption
# - celery and redis only needed for async_queue submission mode
# - gunicorn recommended for production deployments
